add new client conn
