+++ internal/service/shield/service_package.go
@@ -6,21 +6,27 @@ package shield
 import (
 	"context"
 
-	aws_sdkv1 "github.com/aws/aws-sdk-go/aws"
-	endpoints_sdkv1 "github.com/aws/aws-sdk-go/aws/endpoints"
-	session_sdkv1 "github.com/aws/aws-sdk-go/aws/session"
-	shield_sdkv1 "github.com/aws/aws-sdk-go/service/shield"
+	aws_sdkv2 "github.com/aws/aws-sdk-go-v2/aws"
+	shield_sdkv2 "github.com/aws/aws-sdk-go-v2/service/shield"
 )
 
-// NewConn returns a new AWS SDK for Go v1 client for this service package's AWS API.
-func (p *servicePackage) NewConn(ctx context.Context, m map[string]any) (*shield_sdkv1.Shield, error) {
-	sess := m["session"].(*session_sdkv1.Session)
-	config := &aws_sdkv1.Config{Endpoint: aws_sdkv1.String(m["endpoint"].(string))}
+const (
+	awsPartitionID  = "aws"
+	usEast1RegionID = "us-east-1"
+)
+
+// NewClient returns a new AWS SDK for Go v2 client for this service package's AWS API.
+func (p *servicePackage) NewClient(ctx context.Context, config map[string]any) (*shield_sdkv2.Client, error) {
+	cfg := *(config["aws_sdkv2_config"].(*aws_sdkv2.Config))
 
 	// Force "global" services to correct Regions.
-	if m["partition"].(string) == endpoints_sdkv1.AwsPartitionID {
-		config.Region = aws_sdkv1.String(endpoints_sdkv1.UsEast1RegionID)
+	if config["partition"].(string) == awsPartitionID {
+		cfg.Region = usEast1RegionID
 	}
 
-	return shield_sdkv1.New(sess.Copy(config)), nil
+	return shield_sdkv2.NewFromConfig(cfg, func(o *shield_sdkv2.Options) {
+		if endpoint := config["endpoint"].(string); endpoint != "" {
+			o.BaseEndpoint = aws_sdkv2.String(endpoint)
+		}
+	}), nil
 }
